requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
outline-vpn-api @ git+https://github.com/jadolg/outline-vpn-api.git@86db97b3b0e8bc9f8171b2d690db1f7123301e5e
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
